+++ src/controllers/cartridgeCategoryController.ts
@@ -1,9 +1,13 @@
 import type { Request, Response } from "express";
 import { HttpError } from "../utils/errors";
 import * as categoryService from "../services/cartridgeCategoryService";
+import { validate as isUUID } from "uuid";
 
 export async function createCategory(req: Request, res: Response) {
   try {
+    if (!req.body.name) {
+      return res.status(400).json({ message: "Missing required field: name" });
+    }
     const data = await categoryService.createCategory(req.body);
     return res
       .status(201)
@@ -18,6 +22,7 @@ export async function createCategory(req: Request, res: Response) {
 
 export async function getAllCategories(req: Request, res: Response) {
   try {
+
     const data = await categoryService.getAllCategories();
     return res.status(200).json(data);
   } catch (err: any) {
@@ -28,6 +33,12 @@ export async function getAllCategories(req: Request, res: Response) {
 
 export async function getCategoryById(req: Request, res: Response) {
   try {
+    if (!req.params.id) {
+      return res.status(400).json({ message: "Missing required parameter: id" });
+    }
+    if (!isUUID(req.params.id)) {
+      return res.status(400).json({ message: "Invalid ID format. ID must be a valid UUID." });
+    }
     const data = await categoryService.getCategoryById(req.params.id as string);
     return res.status(200).json(data);
   } catch (err: any) {
@@ -41,6 +52,19 @@ export async function getCategoryById(req: Request, res: Response) {
 export async function updateCategory(req: Request, res: Response) {
   try {
     const input = { id: req.params.id, ...req.body };
+   
+
+    if (!input.id) {
+      return res.status(400).json({ message: "Missing required parameter: id" });
+    } else if (!input.name && typeof input.is_active !== "boolean") {
+      return res
+        .status(400)
+        .json({ message: "At least one field (name or is_active) is required for update." });
+    }
+
+     if (!isUUID(req.params.id)) {
+      return res.status(400).json({ message: "Invalid ID format. ID must be a valid UUID." });
+    }
     const data = await categoryService.updateCategory(input);
     return res
       .status(200)
@@ -55,6 +79,12 @@ export async function updateCategory(req: Request, res: Response) {
 
 export async function deleteCategory(req: Request, res: Response) {
   try {
+    if (!req.params.id) {
+      return res.status(400).json({ message: "Missing required parameter: id" });
+    }
+      if (!isUUID(req.params.id)) { 
+        return res.status(400).json({ message: "Invalid ID format. ID must be a valid UUID." });
+      }
     const data = await categoryService.deleteCategory(req.params.id as string);
     return res.status(200).json(data);
   } catch (err: any) {
